+++ brainunit/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 # ==============================================================================
 
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 
 from . import _matplotlib_compat
 from . import lax
